+++ Cult of the Lamb/GlyphOverride/GlobalUsings.cs
@@ -5,5 +5,7 @@ global using BepInEx;
 global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using HarmonyLib;
+global using MMTools;
 global using Rewired;
 global using Shared;
+global using UnityEngine;